moving adding, mylist and removing movies from movie router to user router
