+++ backend/src/TMDBApi.js
@@ -5,7 +5,6 @@ const API_KEY = "29de87d3f58e703ce82ba34e2460edcd"
 
 class TMDBApi {
 
-
     static formatPramaters(params) {
         let formatedParams = ''
         for (let [param, value] of Object.entries(params)) {
@@ -22,6 +21,7 @@ class TMDBApi {
             return undefined
         })
     }
+
     static async movieDetails(movieId) {
         const url = `${API_URL}movie/${movieId}?api_key=${API_KEY}`
         return await axios.get(url).then(res => {
