+++ backend/src/models/Movie.js
@@ -1,6 +1,7 @@
 const mongoose = require("mongoose")
+const Schema = mongoose.Schema;
 
-const movieSchema = mongoose.Schema({
+const movieSchema = new Schema({
     id: {
         type: Number,
         required: true
@@ -22,7 +23,6 @@ const movieSchema = mongoose.Schema({
         required: true
     }
 })
-
 const Movie = mongoose.model("Movie", movieSchema)
 
 module.exports = Movie
