+++ backend/src/routers/movies.js
@@ -1,10 +1,7 @@
 const express = require("express")
 const router = new express.Router()
-const Movie = require("../models/Movie")
 const Helper = require("../Helper")
-const { ObjectID } = require("mongodb")
 const TMDBApi = require("../TMDBApi")
-const auth = require("../middleWare/auth")
 
 // searching for a movie
 router.get("/myFlex/api/v1/search/movie", async (req, res) => {
@@ -17,59 +14,6 @@ router.get("/myFlex/api/v1/search/movie", async (req, res) => {
     }
 })
 
-// adding a movie to your list 
-router.post("/myFlex/api/v1/list", auth, async (req, res) => {
-    try {
 
-        const user = req.user
-        const movieExists = await Movie.findOne({ id: req.body.id })
-
-        if (movieExists === null) {
-            const movie = await TMDBApi.movieDetails(req.body.id)
-            const formatedMovie = Helper.formatMovie(movie)
-            const movieID = new ObjectID()
-            const mongoMovie = new Movie({ ...formatedMovie, _id: movieID })
-            await mongoMovie.save()
-            user.movies = user.movies.concat({ _id: movieID, TMDB_Id: req.body.id })
-        } else {
-            user.movies = user.movies.concat({ _id: movieExists._id, TMDB_Id: req.body.id })
-        }
-        user.save()
-        res.send("Movie Added Succecfully")
-    } catch (err) {
-        res.status(400).send(err)
-    }
-})
-
-// getting my movies list
-router.get("/myFlex/api/v1/list", auth, async (req, res) => {
-    try {
-        const user = req.user
-        const moviesIDList = []
-
-        for (let i = 0; i < user.movies.length; i++) {
-            moviesIDList.push(user.movies[i]._id.toString())
-        }
-        const movies = await Movie.find({ _id: { $in: moviesIDList } })
-
-        res.send(movies)
-    } catch (err) {
-        res.status(400).send(err)
-    }
-})
-
-
-//removing a movie 
-router.delete("/myFlex/api/v1/list", auth, async (req, res) => {
-    try {
-        const user = req.user
-        const movieIndex = user.movies.findIndex(movie => movie.TMDB_Id === req.body.id)
-        user.movies.splice(movieIndex, 1)
-        user.save()
-        res.send("Movie removed Succecfully")
-    } catch (e) {
-        res.status(400).send(err)
-    }
-})
 
 module.exports = router
